+++ cmsafvis/R/monitor_climate.R
@@ -53,6 +53,7 @@
 #' If attach is false, this will be ignored(character).
 #' @param analyze_method Two analyze methods: mean == FALSE; accumulate == TRUE (logical) (Warming Stripes Plots, Time Series Plots, Trend Plots)
 #' @param selected_number Timesteps of the same selected time range (Warming Stripes Plots, Time Series Plots, Trend Plots)
+#' @param stripe_color Color option for stripe plots (Warming Stripes Plots)
 #' @param dwd_logo Whether to add the DWD logo (logical).
 #' @param verbose Whether to display progress messages (logical).
 #' @param nc Alternatively to \code{infile} you can specify the input as an
@@ -94,6 +95,7 @@ monitor_climate <- function(plot_type = "absolute_map",
                             infile_attach = "auto",
                             analyze_method = TRUE,
                             selected_number = 1,
+                            stripe_color = 1,
                             dwd_logo = FALSE,
                             verbose = TRUE,
                             nc = NULL